+++ frontend/src/components/App/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import {BrowserRouter as Router, Route} from 'react-router-dom';
-import { AuthProvider } from '../Auth';
-import PrivateRoute from '../PrivateRoute';
+import { AuthProvider, PrivateRoute } from '../Auth';
 
 import LandingPage from '../Landing';
 import SignInPage from '../SignIn'
